# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACR - Attack Chain Reconstruction Engine",
    "description": "Attack Chain Reconstruction Engine: ingest telemetry, run rule-based detection, correlate events with multi-signal scoring, reconstruct attack chains, map them to MITRE ATT&CK and evaluate against synthetic ground truth.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chains/{chain_id}/investigation": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Investigation",
        "operationId": "chain_investigation_api_chains__chain_id__investigation_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          },
          {
            "name": "include_inferred",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inferred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mitre/techniques": {
      "get": {
        "tags": [
          "mitre"
        ],
        "summary": "List Techniques",
        "operationId": "list_techniques_api_mitre_techniques_get",
        "parameters": [
          {
            "name": "tactic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tactic name",
              "title": "Tactic"
            },
            "description": "Filter by tactic name"
          },
          {
            "name": "include_subtechniques",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Subtechniques"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechniqueListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mitre/techniques/{technique_id}": {
      "get": {
        "tags": [
          "mitre"
        ],
        "summary": "Get Technique",
        "operationId": "get_technique_api_mitre_techniques__technique_id__get",
        "parameters": [
          {
            "name": "technique_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Technique Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechniqueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/mitre/coverage": {
      "get": {
        "tags": [
          "mitre"
        ],
        "summary": "Coverage",
        "description": "ATT&CK coverage of reconstructed chains (one chain, one scenario, or all).",
        "operationId": "coverage_api_mitre_coverage_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chain Id"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoverageOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/investigation/{chain_id}": {
      "get": {
        "tags": [
          "investigation"
        ],
        "summary": "Investigate",
        "operationId": "investigate_api_investigation__chain_id__get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          },
          {
            "name": "include_inferred",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inferred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigation/{chain_id}/summary": {
      "get": {
        "tags": [
          "investigation"
        ],
        "summary": "Investigation Summary",
        "operationId": "investigation_summary_api_investigation__chain_id__summary_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Investigation Summary Api Investigation  Chain Id  Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CoverageOut": {
        "properties": {
          "tactics": {
            "items": {
              "$ref": "#/components/schemas/CoverageTactic"
            },
            "type": "array",
            "title": "Tactics"
          },
          "unique_techniques": {
            "type": "integer",
            "title": "Unique Techniques"
          },
          "unique_tactics": {
            "type": "integer",
            "title": "Unique Tactics"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "chain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Id"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          }
        },
        "type": "object",
        "required": [
          "tactics",
          "unique_techniques",
          "unique_tactics"
        ],
        "title": "CoverageOut"
      },
      "CoverageTactic": {
        "properties": {
          "tactic": {
            "type": "string",
            "title": "Tactic"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/CoverageTechnique"
            },
            "type": "array",
            "title": "Techniques"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "tactic",
          "techniques",
          "count"
        ],
        "title": "CoverageTactic"
      },
      "CoverageTechnique": {
        "properties": {
          "technique_id": {
            "type": "string",
            "title": "Technique Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "technique_id",
          "name"
        ],
        "title": "CoverageTechnique"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationOut": {
        "properties": {
          "chain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Id"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Timeline"
          },
          "process_tree": {
            "items": {
              "$ref": "#/components/schemas/ProcessTreeNode"
            },
            "type": "array",
            "title": "Process Tree"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkViewOut"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "timeline",
          "process_tree",
          "network",
          "summary"
        ],
        "title": "InvestigationOut"
      },
...
      "NetworkViewOut": {
        "properties": {
          "connections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Connections"
          },
          "dns_queries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Dns Queries"
          },
          "external_destinations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "External Destinations"
          },
          "graph": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "title": "NetworkViewOut"
      },
      "ProcessTreeNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "process_name": {
            "type": "string",
            "title": "Process Name"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "ppid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ppid"
          },
          "parent_process_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Process Name"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "command_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Line"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "technique_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/ProcessTreeNode"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "process_name"
        ],
        "title": "ProcessTreeNode"
      },
...
      "TechniqueListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TechniqueOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "tactic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tactic"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "TechniqueListOut"
      },
      "TechniqueOut": {
        "properties": {
          "technique_id": {
            "type": "string",
            "title": "Technique Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "tactic": {
            "type": "string",
            "title": "Tactic"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "subtechnique_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtechnique Of"
          },
          "detection_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Hint"
          },
          "is_subtechnique": {
            "type": "boolean",
            "title": "Is Subtechnique",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "technique_id",
          "name",
          "tactic"
        ],
        "title": "TechniqueOut"
      },
...
      "TimelineItem": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "OBSERVED"
          },
          "inferred": {
            "type": "boolean",
            "title": "Inferred",
            "default": false
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "event_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "command_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Line"
          },
          "entity": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity"
          },
          "relationships": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Relationships"
          },
          "technique": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique"
          },
          "tactic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tactic"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "seq"
        ],
        "title": "TimelineItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}